gh-3163 Dali iterateFiles/ldap optimizations
With LDAP on iterateFiles was authenticating too many possible scopes, not
just those that had matches underneath. There was also a problem after
authentification, where it got out of sync. with the list of authenticated
scopes and repeatedly scanned the list for next match, slowing the
serialization down significantly.

The main change here though is to reduce the # of scopes that need
authenticating against LDAP

Fixes gh-3163

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>