gh-3172 Avoid spurious error on empty inline wildcard superfiles
Creating a temp super that has no matches, causes a spurious error:
"Wildcards not allowed in filename.."

This is because, it doesn't return a multi logical file construct, and
leaves the logical filename with wildcards intact, which is later complains
about.
Fix by creating a multi lfn construct in wildcard case with 0 matches.

Fixes gh-3172

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>