gh-3005 Fix ECLAgent::getPlatform() problem
Currently ICodeContext::getPlatform in eclagent returns "hthor". However
it needs to behave more like getNodes() which looks up the cluster and
returns thor/thorlcr depending on the cluster type. This fix makes calls
to get the cluster name, queries for that cluster's info , and maps the
"platform" attribute to a cluster type string

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>