gh-3203 Script error on ECLWatch from within ECL IDE
When despraying files from within ECL IDE, the following script
error shows: Unable to get value of the proprety 'options'. The
problem was because of an incorrect access to the 'options'
attribute of the 'DestGroup' select box which does not exist on
the Despray form. (The same xslt script is shared by Despray
and Copy). This fix checks existence of select box before
accessing its attribute.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>