gh-658 Add user option to eclplus
Previously the meaning of the owner option was overloaded to mean both
user and the owner to filter on when listing workunits, etc.

Add support for a user option to explicitly specify the user running
the command.

If user is not specified owner will still be used as the username for
backward compatability.

Fixes gh-658.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>