gh-3157 Update query publishing and ws_ecl to use target as queryset name
Change ws_workunits methods WUPublish, WUQuerysets, and *QuerysetDetails
to use the target name and list of targets when dealing with queries
and querysets.

Change ws_ecl to use target as queryest and to list targets in its
treeview, simplifying users understanding of the target queryset.

Change eclwatch to list targets instead of querysets when browsing queries.

Fixes gh-3157.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>