gh-3274 Add proxies for EE-only plugins to CE builds
In order for users of EE to be able to do local syntax-checks / archive
creation on systems with CE client tools installed (and in order to remove
the need for multiple versions of the client tools packages), provide 'proxy'
plugins for the EE-only plugins and include them in CE builds.

Fixes gh-3274.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>