HPCC-2122 Add ability to pause DFU queue
Operations is requesting the ability to pause the DFU queue. The
existing code can pause the queue. But, it still shows that the
queue is running. The existing C++ code sends the queue status
using the first workunit inside the queue. If the queue does not
have a workunit, the UI code cannot know the status. The new code
creates a ServerJobQueue object and uses it to pass the queue
status to xslt.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>