HPCC-3078 Add preflight option to display usage of mounted disks
The existing Preflight function does not check (and not display)
the usage of mounted disks. In this fix, preflight script and
EclWatch UI are upgraded, which allows a user to specify whether
the usage of mounted disks is checked or not.

This fix also cleans two compiler warnings in this ESP service.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>