gh-3321 Fix memory leak for large blocks in memory manager
The heapHWM variable was being updated incorrectly when a block of
memory was allocated.  It should have only been updated if this could
be guaranteed to be the top block of free memory.  Instead it was
updated unconditionally which meant that blocks of memory could be
leaked and could never be reached again.

Experimentation indicated that it was unlikely for the correct update
condition to be met, so it was simplest to remove the heapHWM
altogether, and always search the full memory range.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>