HPCC-3206 Display Delete access denied when no access to a scope
Existing code displays 'Create access denied' when a user deletes a file
but does not have an access to a scope. The message is generated when
EclWatch does a file lookup in order to check whether the file is a
superfile or not. This fix detects the message and converts the message
to 'Delete access denied'.

This fix also improves the CWsDfuEx::DFUDeleteFile() which converts the
error message.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>