HPCC-3116 Deschedule WU even if the WU is not found
A user deleted a workunit where the workunit was in the WAIT state. In
the existing code, the workunit cannot be descheduled because the
deschedule() is a member of IWorkunit and the IWorkunit cannot be opened
after the workunit is deleted. In this fix, an extern descheduleWorkunit
is added to deschedule a workunit. Before the descheduleWorkunit is
called, EclWatch verifies that the user has SecAccess_Full right to both
OWN_WU_ACCESS and OTHERS_WU_ACCESS.

This fix also revises the doAction() function which contains the
deschedule code.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>