HPCC-7820 Add LDAP scope for the per-user temp file location
Temporary files need to be differentiated by user, with separate scopes
restricting access to another user's files. This fix queries DFU for a
tempFiles scope name (HpccInternal), and creates a user-specific scope
(HpccInternal::user) whenever a new user is added. Permissions for that
scope are granted to Administrators and that user only. That scope is
deleted when the user account is deleted.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>