HPCC-3324 ESP can crash if multiple queries received together
If the first time ESP needs to lookup query dlls, it does so on several
threads at the same time, ESP may core due to the global dllServer
object being created multiple times.

Add a CriticalSection to protect the dllServer object. Also copy
strings rather than pointers, in case the lifetime of the dllServer
is ever legitimately shorter than the lifetime of the objects it creates.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>