HPCC-7834 Temporarily dump stack trace if default dali user
In order to help trace the source of the spurious usage of the default
"daliuser" credentials, it would be helpful to (temporarily) log a stack
trace of how we get to that point, hoping to narrow in on the offending
code.
This commit introduces checks in key areas, and logs the stack when the
user is NULL or "daliuser".  The intent is that this code be removed in
subsequent builds when the problem is resolved.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>