HPCC-3198 Remove 'Run ECL' link in EclWatch
The 'Run ECL' link is removed from EclWatch since its functionality
has been replaced by 'ECL Playground'. The EclDirect web service is
still available for JDBC client. This fix also allows to set
username as the job owner (submitID) in the web service. If the
username is not set, the job owner will be blank.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>