HPCC-716 Include superfile into addSuper/removeSuper transactions
When files are added into new superfile and a failure occurs when adding
one of the files, the empty superfile is not deleted. That may mislead
us. This fix moves the superfile creation into the same transaction of
adding the files into the superfile. If there is a failure when adding
the files, the whole transaction will be rolled back so that the empty
superfile will be deleted.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>