HPCC-7868 PackageSet needs a unique id
In order to be able to refer to a particular PackageSet in ecl-package,
we need to give the PackageSet a unique id. Move the name of the roxies
to apply to onto a new process attribute.

Note that this area has already changed since 3.8.x, so this does not
create any (new) incompatibility.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>