HPCC-7870 Convert 'Slave Number' input to integer before comparison
On Ecl Workunit Details page (a system with 100 slave nodes), a user
specifies 39 as a Slave Number and clicks the 'get slave log' button.
An error message shows: 'Slave Number cannot be greater than 100'.
The problem was because javascript identifies the user input using
string and do string comparison. This fix converts 'Slave Number' input
(and total number of slaves) to integer before a comparison.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>