HPCC-7897 Use ReadWriteLock to reduce impact of lock on query lookup table
The lock in CRoxieQueryPackageSetManager will prevent query resolution while
it is held, but is held by the processing of control: messages which may take
a little time.

Switching to a ReadWriteLock will reduce the scope of the problem.

A full solution will require work to reduce the size of the window inside
reload that needs to be locked, but that is outside the scope of this
sub-task.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>