HPCC-559 Set time out value for deleting logical file
When a file cannot be deleted because it is locked, the existing code
blocks the process until SDS connection reaches a default time out.
The default time out is 2 hours. This fix allows a time out value (15
seconds) being set in ESP Ws_dfu service. If the the SDS connection
reaches the time out value. an exception will be thrown and returned
by the ESP service.

This fix changes several daft/dfdfs methods in order to pass the time
out value to the file operation. If the time out value is not set,
those methods uses the default time out value (INFINITE) which is used
before this fix.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>