HPCC-7890 Default to outputting multiple c++ files
Gcc (and other compilers) have problems on very large queries (e.g., >10,000)
activities, because of the total size of a single c++ file.  This changes the
default to split into multiple c++ files.  The regression suite still
generates single files in most cases since that is much easier to compare.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>