HPCC-7919 Report errors that prevented loading a Roxie query
In order to present better feedback to users when queries don't deploy
successfully, improve the output of control:queries to return any
associated error message if a query is suspended. Also allow the output
to be limited to a specified set of queries

The incoming message is of the form:

<control:queries>
 <Query id='queryname'/>
 ... more queries here if you want ...
</control:getQueryInfo>

Code has been refactored to common up the functionality with the just-added
control:getQueryXrefInfo. As a result the output from that query now uses
<Queries> rather than <QueryInfo> as the surrounding element name.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>