HPCC-7852 Display an error when a query is published and suspended
A user publishes a workunit to roxie. The query may be suspended due to
some reason. ESP should be able to notify the user about the problem
when the workunit is published. This fix calls a roxie control command
to check whether a query is suspended or not and, if suspended, displays
the error returned by that command.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>