HPCC-3327 Row arrays limited to 500 million rows
In places (in thor) where an expanding array of rows is allocated from the
roxie/thor memory manager (i.e.. most activities that need to gather large
numbers of rows), we will hit issues (either incorrect behaviour or failures)
when the array size exceeds 4Gb.

With 64-bit pointers, that effectively creates a limit on the number of rows
at 500 million.

Change roxiemem layer to support allocations of > 4gb. This should push the
limit up to 4 billion rows.

We will need to address that limit too, but in a separate issue.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>