HPCC-7965 Set correct content type for xml/xslt to avoid log failure
When ESP loglevel is high and response xml message is logged, the xml
is logged as a SOAP message because the content type is set as 'text/
xml'. This fix changes the content type of xml/xsl/xslt to be
'application/xml'. This fix also adds exception check in to the log
function. When an exception is thrown inside the log function, the
exception and the xml message will be logged into esp.xml.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>