HPCC-7996 Introduce memory manager spill threshold option
Add a thor option to allow the memory manager to be configured, so that it
can spill when it hits a certain threshold(%), default 80%.
The reasoning being, that if the memory manager waits until it is out of
memory before requesting memory to be freed, those requests can take some
time (e.g. sort+flush to disk), and cause all other small memory consumers
to block in the meantime.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>