HPCC-7987 Display getQueryMainDefinition() in ECL WU Details page
It is possible for a query to be submitted with no query text, just a
reference to an attribute in a remote repository. This fix displays
the value of getQueryMainDefinition() in the Query section (inside ECL
WU Details page). The value will be under the textarea of the query
text if there is query text.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>