HPCC-8113 Crash in thor when sorting very large arrays
The changes to roxiemem to support row arrays > 4G caused the aligmnent of
data returned by the huge allocator to change. Previously we got lucky and
it was 8-byte aligned, but after the change it was 4 byte aligned.

Cleaned up the code so that the same alignment fix that was applied when we
hit this issue on small arrays applies to big ones too.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>