HPCC-8064 Roxie passing NULL for IUserDescriptor
If no username and password has been supplied in the topology file, Roxie
currently passes NULL to indicate the default user should be used. However,
as we are trying to get rid of all uses if the defualt user, and tracing
aggressively when they are seen, change Roxie to use the userid "roxie"
and blank password if none is supplied.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>