HPCC-8103 Fix missing hyperlinks for results in WU Details page
The existing WU Details page uses the ShowFileContent flag to decide
whether the Result hyperlinks should be displayed or not and the flag
is set to true as default. From 3.10.x, the default value is not set
for HTTP response and the Result hyperlink may not be displayed.

In fact, the ShowFileContent flag is being set only if the result has
a logical file name. If no file name, the WU Details page should not
check the ShowFileContent flag.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>