HPCC-8098 Fix seh after unusual syntax error
Currently pushSelf is used differently when processing a record from
when processing a transform.  In the former it is a record, in the latter
a no_self.  This should really be changed in the future and made consistent.

A user had an example query containing a syntax error which caused the
state to become inconsistent, and then crash.  This change fixes the seh by
ensuring the record of the active item is selected.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>