HPCC-7980 Translate remote file dafilesrv access to local
Add a mechanism to spot remote file access that matches a/some
subnets and translate them to local file access.

This is allow separate slaves/nodes to circumvent the use
of dafilesrv access remotely on other nodes, when the FS is always
remote, e.g. a NAS fs attached to all nodes

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>