HPCC-8102 Rename eclccserver maxCompileThreads option
To avoid confusion with the similarly-named option in eclcc, that does
something slightly different, rename the eclccserver option to
maxEclccProcesses.

The old option name will still be accepted if the new is not present.

Also make the default value consistent between the xsd file and the cpp
file - it now defaults to 4 in both places, where previously the cpp
would default to 1 if the value was explicitly deleted from the config
(unlikely).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>