HPCC-8176 Regression in library resolve code
The rewriten code for resolving libraries was flawed, in that each query
kept a pointer to an object that would be destroyed by the next query load
operation.

The selection of the query manager to use for library resolution needs to
be delayed until actual library resolution time.

This code fixes the core, but requires that the query use the same package
as any library it uses. It's not clear if that is a sensible restriction
without some more thought.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>