HPCC-7970 Include eclcc timing information in workunits
The workunit embedded in the .so file contains timing info, but it was
not copied into the dali workunit by eclccserver.

Also, some interesting timing (like the time to compile the c++) does
not get captured because the workunit has already been dumped to XML
so that it can be included in the .so - capture this info via a
different mechanism.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>