HPCC-8185 Protect against corrupt or truncated end of dali delta
If there is a cold boot mid write, the end of the transaction
delta can have incomplete XML and cause problems when Dali
restarts. To help prevent this, write the last know commited file
size to the header (together with crc).
Use this known 'good size' to verify the file doesn't have
trailing unknown data.
If the file mismatches, report the problem, backup the file
and truncate to the known committed size and continue.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>