HPCC-8215 Regenerate HEADING correctly instead of as HEADER
The wrong atom name was being used when the HEADING attribute was
represented in the expression tree which meant it was displayed as
HEADER.  This uses the correct name.

The change has the side-effect of modifying the eclcrc for files with
HEADING which means that OUTPUT,UPDATE will trigger a rebuild.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>