HPCC-8236 Avoid spilling callback deadlock
A race condition in Thor's use of the spilling callback mechanism,
could mean that it deadlocks. A call back occurs whilst it is
preparing to finish and setup a stream of existing spills, at
that point, it sets up another callback, but that will get
blocked whilst in the existing callback = deadlock.

Fix is to clear the existing callback which is finished with,
before entering the lock and setting up the new one.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>