HPCC-8250 Activities derived from disk read asserting on multiway files
When reading a multiway file, some activities derived from diskread are
failing with assertion errors due to changes related to the delayed
decision on whether remote read is needed.

Now that the queryOutput never returns the remote adaptor, the redirect
to use remote needs to be added to the nextInGroup methods for csv, xml
and normalize activities.

In addition, where OPT is used and a file ends up with no data, some
fields used by the local read code were not set up.

This error was introduced in rc3 and should be addressed before rc4/gold.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>