HPCC-8026 Fix View Log file in Last N Hours option
If a log file is large, the code of View Log file will check
timestamp from the last chuck of the log. The Last N Hours option
fails due to incorrectly match quotes when the chuck starts
from the middle of a log line and an open quote is not inside the
chuck. (Quotes may be used to hold a big log line which are broken
into multiple physical log lines). This fix is based on the
following assumption: if the log line contains a timestamp and the
timestamp is at expected location, the log line is a standard log
line; if not, it is a continuation line.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>