HPCC-8088 Hide / collapse ftslave in System Servers
We have several environments with 1000+ servers in them. By default
ftslave isn't something that we preflight. It would be nice if the
ftslaves were hidden or collapsed by default. This fix hides the
ftslaves inside a folder when the System Servers page is loaded.
A user may open the folder as needed.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>