HPCC-8183 Display a better message when uploading file fails
The existing code does not do anything (just breaks the HTTP
connect) if a dropzone does not have access right for uploading
file. As a result, the connection reset error is displayed on
browser. This fix catches all of exceptions for uploading file
and display the exception message in a html page.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>