HPCC-3208 Report a warning if a scalar output is used in an APPLY
It doesn't make sense to use a scalar output within an apply since a
scalar output can only output a single value.
It also causse Thor to generate an internal error since the output code is
executed on the slave which Thor doesn't support.

It isn't made an error because there are some existing queries which use it.
(They should be modified.)

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>