HPCC-1766 Use correct paths for non-local cluster writes
Thor always constructed physical file paths based on the local
thor it was running on. That was incorrect if the target cluster
was different, e.g. with OUTPUT,CLUSTER('other')
Most of the time the pathing is the same, so it's a non-issue
It is only exposed when the environment specifies different
directories with e.g. <Override instance="athor" dir="a/b/c"/>

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>