HPCC-8358 ESP Soap arrays of structures not xml decoding input strings
The flag that tells the ESP SOAP interface handlers not to XML encode
incoming string values needs to be passed down to the child elements
of structure members when they are held in arrays.

This change should only affect the encoding of data retrieved from
incoming XML/SOAP requests, and only those which contain arrays...
especially nested structures inside of arrays.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>