HPCC-8361 Initial implementation of new IR processor
This commit contains a new approach to the Intermediate Representation
generator.  It is based around a builder interface which is should
be target format agnostic.  This means the same interface can be used
for generating

a) text.
b) xml.
c) binary serialied format.
d) an expression tree.

It should also allow inputs from each of those formats to be "played" and
converted into any of the others.  This will allow us in the future to have
a serialized format for the expression tree, which could be useful for
debugging, serializing to a remote server, or even serializing to the engines
to allow expressions to be interpreted.

Currently the only classes implemented are the player for the expression
trees and the text output.  The binary reader has been sketched out.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>