HPCC-8233 Fix JSONTest page not populating "set of strings"
JSONTest will populate a JSON request from the content of a WsECL
test form.  When the input is a stored set of string, the form data
was not populating the JSON request.

This issue only affects generating the sample request.  It doesn't
affect processing of JSON requests once they are submitted.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>