HPCC-8494 Fix incompatibility in the message format for ftslave
The serialized information that was sent from the master to the slave
changed between 3.8.x and 3.10.0 in a way that was not backward compatible.

This makes it impossible to spray csv (or xml) files from a system on one version
to another.  This change ensures the different versions of the systems can
work together.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>