HPCC-8577 - Unecessarily slow thor start/stop times
The default frunssh behaviour is to limit concurrent threads to 10
This means the starting process needs to wait until some of the
start slave script complete, before beginning to start others.
There is no good reason the limit is 10 and in general it should
start all in parallel, via option to frunssh.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>