HPCC-8610 Ensure all Eclagent logfiles get recorded in workunit
Currently, if a running workunit spans past midnight, the logfile rolls
over to a new daily one, but eclagent never updates the workunit with a
reference to the new one. This fix calls addProcess to record the logfile
after every graph and subgraph execution. Note that if addProcess is called
on the same day (logfile name has not changed) then it will NOT get
recorded again.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>