HPCC-8599 Roxie may core when removing a file from file cache
If a file is on the 'to copy' list at the time that it is removed from the
Roxie file cache, then the reference on the 'to copy' list becomes invalid.

This problem was introduced when the roxie file cache was changed from being
an 'owning' cache to a 'non-owning' cache, so that files were not locked by
roxie for longer than they should be.

There are also some subtelties to use of a 'non-owning' cache where even
inside a critsec, you may retrieve a pointer that is about to be removed from
the cache and freed - this requires the isue of isAlive.

Also removed some unused code that would only have worked in the old owned
cache.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>