HPCC-8413 Fix problems with STEPPED JOIN accessing non existant results
A stepped join tries to prevent the inputs from creating splitters - since that
will default the smart stepping on indexes.  However if the inputs were not
indexesw then it could generate an internal error since one part of the graph
assumed a spill would be created, but none ever was.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>