HPCC-8030 Java/Python/Javascript language support in ECL
Remove SONAME from these 'module' plugins, so that installers don't complain
about missing dependencies. That way they will be available IFF the
corresponding language they interface to is installed, but not require that it
is for the HPCC-Platform to be installed.

Also clean up the build process to use proper CMake mechanism for finding
includes and libraries, and to add flags to disable building these plugins.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>