HPCC-825 Change Resubmit and Restart buttons to be less confusing
Restart currently does what Resubmit used to do, which was to reset
and restart the workunit from the beginging.  Instead Resubmit now
tries to recover the workunit from where it left off.

First change the Resubmit behavior back to what it was previously.

Also remove "Restart" and add a "Recover" button which will attempt
to begin the workunit from where it previously stopped.

Also improve the tooltip to better explain what will happen.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>