HPCC-8727 ONCE dictionaries should not be serialized to the slaves
Change how and when ONCE sections are handled, so that we can avoid
serializing data from a ONCE section to the slave nodes. This is particularly
useful when a DICTIONARY is createdin a ONCE section (for example by loading
data from a file) but may also apply elsewhere.

Refactored to ensure that the ONCE context is shared between the server and
any slave channels on a node. Also added the option (defaulted on) for ONCE
sections to be evaluated when queries are loaded rather than on first access,
to avoid prolonged 'cache warmup' times when slaves use ONCE data.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>