HPCC-8767 ESP users should be able to reset expired passwords
Currently ESP (LDAP) users have to have an admin reset expired passwords.
This change allows a user to change it themselves. It is ensured that people
cannot change someone else's password by requiring the "old password."
When passing a valid but expired password to LDAP, the authentication
fails but returns a unique error string that is parsed to detect this scenario

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>