HPCC-8726 Fix regression errors on KEYED(key-field in <dictionary>)
On reflection, I don't think auto-translating field IN dictionary to use a
segmonitor set is particularly useful feature, and the existing behaviour is
correct.

Change the test case so that it manually translates to a set, and add a new
one for unkeyed field IN DICTIONARY filtering.

Also correct some other minor issues in the test/key.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>