HPCC-8817 More changes based on new request
The changes in this commit are based on the following suggestions.
1. Activity page should indicate which WU is running on an ECLagent
and which WU is 'really' running on a cluster. For the WU running
on an ECLagent, '(on ECLagent)' will be displayed inside its State
column. 2. The WUs running on an ECLagent should be list after
queued WUs under a cluster.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>