HPCC-8816 Show 'Delete Query' result and redirect to QuerySet
After a query is deleted using the Delete button on QueryDetails page,
the QueryDetails page becomes invalid which should not be displayed.
This fix redirects the page to Browse Queryset page after the query
has been deleted.

On the Query Details page, a check box is added to indicate the query
is activated or not. The check box is also used for activating and
deactivating the query.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>