HPCC-8911 - Deadlock adding duplicate subfile to super
If an existing subfile was added to a super, the transaction
mechanism would cause a locking deadlock on the 2nd add.
The 2nd add would attempt to gain an exclusive lock, having
already lookup the super+sub for read, thereby locking.
The transaction mechanism needs to check that the subfile
is not part of the super before that happens.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>