HPCC-8910 Rename EclWatch Copy URL menu to View URL
The EclWatch Copy URL icon was used to copy current EclWatch URL to
clipboard. Since most of browsers does not allow JavaScript to
access clipboard data for security and privacy reasons, the icon is
used to display the URL to a new window and a user may copy the URL
from that window. This fix changes the popup text of the icon from
'Copy URL' to 'View URL'. This fix also changes the window size
from 400 x 50 to 800 x 500 for a better display.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>