HPCC-8937 Add a flag to skip existing superfile check
EclWatch has a Radio button on Add to Superfile page. The button
allows a user to specify a subfile is added to an existing superfile
or a new superfile. EclWatch will check whether the superfile exists
or not and show an error message if the user choose a wrong radio
button. But, EclWatch should not do the check if addSuper request
comes from dfuplus/addSuper command (not come from the EclWatch page)
because the existing addSuper command does not specify that a subfile
is added to an existing superfile or a new superfile. This fix adds
a flag to avoid the existing superfile check if the addSuper request
comes from dfuplus/addSuper command.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>