HPCC-8761 Add HTTP 304 support to HTTP requests
This fix adds HTTP 304 support to ESP HTTP requests to improve ESP
performance. When an HTTP GetFile request is received, ESP reads
the timestamp of the file. If the request does not contain an etag
which contains the timestamp, ESP will create the etag and send
the etag with the content of the file. Otherwise, ESP will not
read the file content and send an HTTP 304 'Not Modified' response
to the client. If the HTTP request contains a 'if-modified-since'
header without an etag header, ESP will only change whether the
value of 'if-modified-since' matches the file's modified time or
not. If matches, an HTTP 304 response will be sent.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>