HPCC-9011 Check user object before checking its password expiration
The existing code checks a user's password expiration using the
user object in ESP. If no user authentication is required, the user
object is NULL and we do not need to check user status of
authentication. This fix checks the user object before using it to
checks the password expiration.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>