HPCC-9998 Replace 'children' property with 'childNodes' in javascript
Using Firefox 19.0.2, open ECLWatch and select a file, or a user, or a
group, etc and click the DELETE button will result in each selected
item being displayed 4 times. The problem is because that the new version
of Firefox does not support the 'children' property as well as its
previous versions. In an HTML form, when a checkbox is checked, Firefox
19.0.2 lists four items being checked if we use the 'children' property
to go through the form items. This fix replaces the 'children' property
with childNodes property which is supported in all major browsers.

This fix also removes three functions which are not being used any more
(the 'children' property is called in those functions).

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>