HPCC-9068 Embedded scripting plugins need to syntax check without dll
If a plugin cannot be loaded (e.g. because of missing dependencies) then you
would not be able to create archives.

Split the ecl code from the executable for the 'optional' plugin components so
that executable code is not needed at syntax check time.

We may want to consider using the same paradigm for all plugins, at some
point, but no pressing need to do so.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>