HPCC-9108 Solve problem not commoning up selectors
The main change is in hqlttcpp.cpp - checking the location independent
version of the transformed selector, and in hqlattr.cpp removing the
arguments of a no_attr to avoid problems with original attributes not
being transformed.

Other changes are minor bug fixes in hqlattr, and extensions to EclIR to
make it more flexible.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>