HPCC-7999 Invalid syntax reported by OpenLDAP
ESP fails to start on newer releases of openLDAP because of numerous
issues relating to schema changes.  Discussed with Flavio and it was
decided to focus on getting it up and running with Fedora389 instead,
since that was a better open source LDAP implementation. With this fix
I am making several tweaks the allow it to start up, create the top level
organizational units, and perform user/group/permission activities.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>